+++ assets/javascripts/kitten/components/box/link-box.js
@@ -45,16 +45,9 @@ export class LinkBox extends Component {
             <p className="k-LinkBox__text">{ text }</p>
           </div>
 
-          <div className={ classNames('k-LinkBox__navigation',
-                                      'k-LinkBox__navigation--withAnimation') }
+          <div className={ classNames('k-LinkBox__navigation') }
           >
-            <span className={ classNames('k-LinkBox__button',
-                                         'k-ButtonIcon',
-                                         'k-ButtonIcon--hydrogen',
-                                         'k-ButtonIcon--tiny',
-                                         'k-ButtonIcon--withoutHover',
-                                         'k-ButtonIcon--verticalArrow') }
-            >
+            <span>
               <ArrowIcon className="k-ButtonIcon__svg" />
             </span>
           </div>